+++ Cargo.toml
@@ -45,6 +45,7 @@ humantime-serde = "1.0"
 medea-client-api-proto = { path = "proto/client-api", features = ["medea"] }
 medea-control-api-proto = { path = "proto/control-api" }
 medea-macro = { path = "crates/medea-macro" }
+medea-reactive = { path = "crates/medea-reactive" }
 rand = "0.7"
 redis = "0.15"
 rust-crypto = "0.2"
