+++ jason/src/api/room.rs
@@ -896,7 +896,7 @@ impl EventHandler for InnerRoom {
     /// Creates new `Track`s, updates existing [`Sender`]s/[`Receiver`]s with
     /// [`TrackUpdate`]s.
     ///
-    /// Will start negotiation process if `Some` [`NegotiationRole`] is
+    /// Will start (re)negotiation process if `Some` [`NegotiationRole`] is
     /// provided.
     async fn on_tracks_applied(
         &self,
