+++ jason/src/peer/conn.rs
@@ -578,7 +578,7 @@ impl RtcPeerConnection {
     /// description.
     ///
     /// Should be called after local tracks changes, which require
-    /// negotiation.
+    /// (re)negotiation.
     ///
     /// # Errors
     ///
