+++ proto/client-api/src/lib.rs
@@ -296,7 +296,7 @@ pub enum Event {
         /// Negotiation role basing on which should be sent
         /// [`Command::MakeSdpOffer`] or [`Command::MakeSdpAnswer`].
         ///
-        /// If `None` then no negotiation should be done.
+        /// If `None` then no (re)negotiation should be done.
         negotiation_role: Option<NegotiationRole>,
     },
 }
